fixed opcode type was uint64_t instead of uint16_t

+++ opcode_specification.h
@@ -15,7 +15,7 @@ typedef enum {
     // TODO: Difference?
 } ArgumentType;
 
-typedef uint64_t Opcode;
+typedef uint16_t Opcode;
 
 typedef struct {
     StringView name; // e.g. "MoveRegisterImmediate"
